requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.1; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django-ckeditor==6.4.0
django-js-asset==2.0.0; python_version >= '3.6'
django==4.0.4
sqlparse==0.4.2; python_version >= '3.5'
